oops, fix size of integer representation of the score.  issue #65

+++ postgres/src/main/c/am/zdbseqscan.h
@@ -21,7 +21,7 @@
 
 typedef struct {
     union {
-        int8 iscore;
+        int32 iscore;
         float4 fscore;
     };
 } ZDBScore;
